- Changed IngredientHolder amount type to EditText - Close keyboard and hide amount seekbar on "Done" action and release of seekbar - Changed ingredient amount from long to int to set progress of seekbar without casting - Set ingredient amount input field to action "Done" and added seekbar to v26 definition - Increased amount input size

+++ inventorychef/src/main/java/net/d4rk/inventorychef/database/dao/Ingredient.java
@@ -26,7 +26,7 @@ import android.arch.persistence.room.PrimaryKey;
     private String unit;
 
     @ColumnInfo(name = "Amount")
-    private long amount = 0;
+    private int amount = 0;
 
     @ColumnInfo(name = "Deleted")
     private boolean deleted;
@@ -74,11 +74,11 @@ import android.arch.persistence.room.PrimaryKey;
         this.unit = unit;
     }
 
-    public long getAmount() {
+    public int getAmount() {
         return amount;
     }
 
-    public void setAmount(long amount) {
+    public void setAmount(int amount) {
         this.amount = amount;
     }
